First refactoring
First refactoring
